requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==5.0.6
django-cors-headers==4.4.0
django-rest-knox==4.2.0
djangorestframework==3.15.2
idna==3.7
pycparser==2.22
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2